fix gcc 10.3 compiler warnings.

+++ minmea.h
@@ -275,7 +275,7 @@ static inline float minmea_tocoord(const struct minmea_float *f)
         return NAN;
     int_least32_t degrees = f->value / (f->scale * 100);
     int_least32_t minutes = f->value % (f->scale * 100);
-    return (float) degrees + (float) minutes / (60 * f->scale);
+    return (float) degrees + (float) minutes / (60 * (float)f->scale);
 }
 
 /**
